planner-api/src/main/java/com/macalester/mealplanner/validator/PasswordConstraintValidator.java
package com.macalester.mealplanner.validator;

import jakarta.validation.ConstraintValidator;
import jakarta.validation.ConstraintValidatorContext;

public class PasswordConstraintValidator implements ConstraintValidator<PasswordConstraint, String> {
    @Override
    public void initialize(PasswordConstraint constraintAnnotation) {
        ConstraintValidator.super.initialize(constraintAnnotation);
    }

    @Override
    public boolean isValid(String s, ConstraintValidatorContext constraintValidatorContext) {
        if (s == null) {
            constraintValidatorContext.disableDefaultConstraintViolation();
            constraintValidatorContext.buildConstraintViolationWithTemplate("{password.error.null}").addConstraintViolation();
            return false;
        }

        return !(s.isBlank() || !s.matches("^[A-Za-z0-9 !#$%&?@^_]+$"));
    }
}
